Fixed public Wikis displaying when user downgrades
Added some direct SQL to the scope in Wiki.rb to cover the `Nil` and `false` values of `Private` boolean. So basically, the wikis get Nil put in the Private boolean when they downgrade, and then that newly public wiki is included in the Wiki.rb scope because the value is Nil (will also work with false).

Realized that there can be actually 3 vakuse toa boolean - nil, tru or false!